Refactor frontend default base layer configuration
